# api-specs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: Fruition
  description: Fruition PoC 백엔드 API
  title: Fruition API
  version: 0.0.1
security:
- bearerAuth: []
...
paths:
...
  /api/workspaces:
    get:
      description: 로그인한 사용자가 소유한 워크스페이스 목록을 반환합니다.
      operationId: list
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceListResponse"
          description: 조회 성공
      summary: 워크스페이스 목록 조회
      tags:
      - Workspaces
    post:
      description: 로그인한 사용자 소유의 워크스페이스를 생성합니다.
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceCreateRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
          description: 생성 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 잘못된 요청
      summary: 워크스페이스 생성
      tags:
      - Workspaces
...
  /api/workspaces/{workspace_id}:
    delete:
      description: 소유한 워크스페이스를 하위 데이터 변경 없이 소프트 삭제합니다.
      operationId: delete
      parameters:
      - description: 워크스페이스 ID
        example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - example: 550e8400-e29b-41d4-a716-446655440000
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceLifecycleResponse"
          description: 삭제 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 잘못된 Idempotency-Key
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Idempotency-Key 충돌
      summary: 워크스페이스 삭제
      tags:
      - Workspaces
    patch:
      description: 로그인한 사용자가 소유한 워크스페이스의 이름을 변경합니다.
      operationId: rename
      parameters:
      - description: 워크스페이스 ID
        example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceRenameRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
          description: 변경 성공
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
      summary: 워크스페이스 이름 변경
      tags:
      - Workspaces
  /api/workspaces/{workspace_id}/icon:
    put:
      description: 소유한 워크스페이스의 아이콘 이모지를 설정하거나 지웁니다.
      operationId: updateIcon
      parameters:
      - description: 워크스페이스 ID
        example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceIconUpdateRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
          description: 변경 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 잘못된 요청
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
      summary: 워크스페이스 아이콘 변경
      tags:
      - Workspaces
  /api/workspaces/{workspace_id}/icon/image:
    get:
      description: 멤버에게 아이콘 이미지 bytes를 반환합니다.
      operationId: getIconImage
      parameters:
      - description: 워크스페이스 ID
        example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: string
                format: byte
          description: 이미지 반환
        "304":
          content:
            '*/*':
              schema:
                type: string
                format: byte
          description: 캐시된 이미지 사용
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스 또는 아이콘 이미지를 찾을 수 없음
      summary: 워크스페이스 아이콘 이미지 조회
      tags:
      - Workspaces
    put:
      description: 아이콘으로 쓸 이미지를 업로드합니다. 이모지가 설정돼 있었다면 함께 사라집니다.
      operationId: updateIconImage
      parameters:
      - description: 워크스페이스 ID
        example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
          description: 업로드 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 지원하지 않는 이미지 형식
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 워크스페이스를 찾을 수 없음
        "413":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 이미지가 너무 큼
      summary: 워크스페이스 아이콘 이미지 업로드
      tags:
      - Workspaces
...
  /api/workspaces/{workspace_id}/restore:
    post:
      description: 소프트 삭제한 워크스페이스와 기존 하위 데이터의 접근을 복구합니다.
      operationId: restore
      parameters:
      - example: ws_9d47a0e9a6324341b47562553b75f92a
        in: path
        name: workspace_id
        required: true
        schema:
          type: string
          default: ws_9d47a0e9a6324341b47562553b75f92a
      - example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
        in: header
        name: Idempotency-Key
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkspaceLifecycleResponse"
          description: 복구 성공
        "400":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 잘못된 Idempotency-Key
        "404":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 삭제 workspace 또는 소유권을 찾을 수 없음
        "409":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Idempotency-Key 충돌
        "410":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: 휴지통 보관 기간이 지나 복구할 수 없음(WORKSPACE_TRASH_EXPIRED)
      summary: 워크스페이스 복구
      tags:
      - Workspaces
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 분기 기준으로 쓰는 값. message가 아니라 이 값으로 분기한다.
          example: INVALID_REQUEST
        details:
          type: array
          description: 검증 실패(400 INVALID_REQUEST)일 때만 존재한다. 값이 없으면 키 자체가 빠지므로 null 체크가 아니라 존재 여부로 확인한다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: 사용자에게 그대로 보여도 되는 한국어 문구
          example: 요청 형식이 올바르지 않습니다.
    ErrorResponse:
      type: object
      description: 컨트롤러까지 도달한 요청이 실패했을 때 모든 API가 공유하는 에러 응답 형식. 인증 필터가 막는 401은 HttpStatusEntryPoint가 상태 코드만 내보내므로 이 본문이 없다.
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
          description: 에러 상세
    FieldError:
      type: object
      properties:
        field:
          type: string
          description: 문제가 된 요청 필드명
          example: email
        reason:
          type: string
          description: 그 필드의 위반 사유
          example: email은 필수입니다.
...
    WorkspaceCreateRequest:
      type: object
      properties:
        name:
          type: string
          description: 워크스페이스 이름(255자 이하). 비워 두면 "새 워크스페이스"가 된다. 이미 같은 이름이 있으면 뒤에 번호가 붙는다.
          example: 내 워크스페이스
          maxLength: 255
          minLength: 0
    WorkspaceIconUpdateRequest:
      type: object
      properties:
        icon_emoji:
          type: string
          description: 아이콘으로 쓸 이모지. null을 주면 아이콘을 지운다.
          example: 📁
          maxLength: 32
          minLength: 0
          pattern: \S+
...
    WorkspaceLifecycleResponse:
      type: object
      properties:
        deleted:
          type: boolean
          description: 삭제(휴지통) 상태인지 여부. 복구하면 false가 된다.
          example: true
        deleted_at:
          type: string
          format: date-time
          description: 삭제된 시각(ISO-8601 UTC). 복구된 경우 null이다.
          example: 2026-08-13T04:25:24.371948Z
        id:
          type: string
          description: 대상 워크스페이스 ID
          example: ws_9d47a0e9a6324341b47562553b75f92a
    WorkspaceListResponse:
      type: object
      properties:
        workspaces:
          type: array
          description: 내가 활성 멤버인 워크스페이스 목록
          items:
            $ref: "#/components/schemas/WorkspaceResponse"
...
    WorkspaceRenameRequest:
      type: object
      properties:
        name:
          type: string
          description: 새 워크스페이스 이름(255자 이하)
          example: 이름 바꾼 워크스페이스
          maxLength: 255
          minLength: 0
      required:
      - name
    WorkspaceResponse:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          description: 생성 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
        icon_emoji:
          type: string
          description: 아이콘 이모지. 설정하지 않았으면 null이다.
          example: 📁
        icon_url:
          type: string
          description: 아이콘 이미지 조회 경로. 이미지를 설정하지 않았으면 null이다.
          example: /api/workspaces/ws_9d47a0e9a6324341b47562553b75f92a/icon/image
        id:
          type: string
          description: 워크스페이스 ID. 이후 대부분의 API 경로에 들어간다.
          example: ws_9d47a0e9a6324341b47562553b75f92a
        name:
          type: string
          description: 워크스페이스 이름
          example: 내 워크스페이스
        updated_at:
          type: string
          format: date-time
          description: 마지막 변경 시각(ISO-8601 UTC)
          example: 2026-08-13T04:25:24.371948Z
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http